use bitflags for GcHeadFlag

+++ src/mark_sweep.rs
@@ -178,12 +178,15 @@ impl GcHeap {
         // clear weakref
         if let Some(weakref_index) = unsafe { (*node.as_ptr()).weakref_index() } {
             debug_assert!(
-                weakref_index < self.weak_list.len(),
+                (weakref_index as usize) < self.weak_list.len(),
                 "{weakref_index}, {:?}",
                 self.weak_list.len()
             );
             unsafe {
-                self.weak_list.get_unchecked_mut(weakref_index).1.take();
+                self.weak_list
+                    .get_unchecked_mut(weakref_index as usize)
+                    .1
+                    .take();
                 (*node.as_ptr()).set_weakref_index(None);
             }
         }
